initial release: wasta branding for calamares installer
... based off of calamares-settings-lubuntu for 18.04, 19.10.

There is some stuff in 19.10 related to EFI,
but that depends on the newer contextualprocess module.
A fair amount of hacky stuff happening with that.

TODO: something is breaking lightdm from listing the users
on the greeter page. Perhaps this is related to systemctl user@123
(lightdm user) failing pam_authentication after a calamares install.

TODO: unsquashfs is very slow, reportedly because it uses
rsync instead of unsquash. A reference was made to rawfs as a
related solution, or perhaps use unsquash when full disk is erased?
The point is that the installer wants to support replace partition
which selectively copies files out.
  see Switch to using unsquashfs for squashfs-based installs in unpackfs #1118
  https://github.com/calamares/calamares/issues/1118

TODO: it doesn't install from multibootusb. Is there an environment
variable we can prefix to the path?